requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
mysql-connector-python==8.3.0
google-generativeai==0.3.2
sqlparse==0.4.4
pydantic==2.5.3
python-multipart==0.0.6
fastapi-cors==0.0.6
